[TEST HARNESS] Added the cpclient test harness.
The test harness is a small python script that simulates provman.
If you run this script before sending an OMA CP WAP Push message
to your device, it will intercept all the settings intended for
Provman.  The settings are stored in a text file.  The intention
is that this test file can be compared against a correct set of
settings for the OMA CP message that has just been processed, thus
permitting automated testing of the cpclient independently of
Provman.